docker should be an optional dependency
we currently are just using it for login and finding configs, and since
osx on conda does not have it (and we want to support it) it makes sense
to have docker be optional, meaning we fall back to manual ways to
do the same thing. I am leaving it as an extra variant for now in case
someone prefers using docker, however if we have confidence it is not
needed it can be removed entirely.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>