adding 1_python.md site for oras-www (#27)
* try removing pytest-runner
* adding 1_python.md site for oras-www
* fixing -k to be correctly associated with keeping files
what we can do here is add a page here, and then setup a workflow in oras-www to receive a
dispatch event to do a build when it is triggered. I can first test this locally (and it will open a PR)
and when we like that, we can add the dispatch event to be triggered in oras-py, and then push directly
given that the commit is signed. Given that netlify is set up to build on push to main, if this
works it will mean that we always trigger an update and rebuild on oras-www when there is an update to
oras-py

Signed-off-by: vsoch <vsoch@users.noreply.github.com>