refactoring docs to markdown instead of rst
the drawback is that we lose the nice toctree offered by
rst, and instead need to hard code the sidebar in the conf.py
as a variable. This should be OK given that we do not want
to hugely change the structure of the site. We are essentially
choosing user ease of editing markdown over automatically
generated sidebars

Signed-off-by: vsoch <vsoch@users.noreply.github.com>