replacing system exits with raising exceptions
Using the client within python should not exit,
but instead raise an exception for the developer user
to handle. This change removes system exits from the
main API, and preserves them in only the command line
client functions to provide a better user experience
(a cleanly formatted error message instead of a stack
trace).

Signed-off-by: vsoch <vsoch@users.noreply.github.com>