removing python client
As discussed in slack/Twitter, providing an equivalent command
line client in two languages is redundant - a user wanting
a command line client out of the box should use the Go variant
at oras-project/oras. This change refactors Oras python to be
explicitly an SDK, and removes this redundant client. The
previous client has been refactored into simple examples, and
moving forward when a registry has a snowflake setup or flow
that is needed, we can craft an example instead of needing
to provide one client to rule them all.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>