Remove Client (#59)
* namespace is not a hard requirement

this change adjusts the parser to not fail without a namespace,
and allows parsing the URI to not include it if it is not
found in the regex
* pin mypy to 0.961
* add api_prefix with conditional add of namespace
* remove colon from message (can be confusing)
* install types-requests
* auth might not be under hostname key
* add debug printing
* set correct token type
* add more developer and debugging docs
* honor username /password if set in pull/push
* add support for custom media type
* blob can be posixpath
* removing python client
* remove push workflow - no client

As discussed in slack/Twitter, providing an equivalent command
line client in two languages is redundant - a user wanting
a command line client out of the box should use the Go variant
at oras-project/oras. This change refactors Oras python to be
explicitly an SDK, and removes this redundant client. The
previous client has been refactored into simple examples, and
moving forward when a registry has a snowflake setup or flow
that is needed, we can craft an example instead of needing
to provide one client to rule them all.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>