Removing the additional duplicate sub-directory from sanitize_path()
The existing implementation adds an additional sub-directory with the same name when an image is pulled and is stored explicitly to a custom output directory. This happens because the "test_path" variable appends "path" to "base_dir" wherein the "path" variable itself is base_dir + filename. This commit fixes this problem.

Signed-off-by: saketjajoo <saketjajoo77@gmail.com>